выполнил ДЗ номер 13
