--- module12/src/main/java/ru/sberbank/edu/config/MySecurityConfig.java
@@ -1,29 +1,0 @@
-package ru.sberbank.edu.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.crypto.password.NoOpPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Configuration
-@EnableWebSecurity
-public class MySecurityConfig extends WebSecurityConfigurerAdapter {
-
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        http.authorizeRequests()
-                .antMatchers("/users").permitAll()
-                .antMatchers("/users/*").permitAll()
-                .anyRequest().authenticated()
-                .and()
-                .formLogin();
-    }
-
-    @Bean
-    public PasswordEncoder passwordEncoder () {
-        return NoOpPasswordEncoder.getInstance();
-    }
-}
