--- module12/src/main/java/ru/sberbank/edu/entity/User.java
@@ -1,51 +1,0 @@
-package ru.sberbank.edu.entity;
-
-import lombok.*;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "users")
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private Long id;
-
-    private String name;
-    private Integer age;
-
-    public User(String name, Integer age) {
-        this.name = name;
-        this.age = age;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Integer getAge() {
-        return age;
-    }
-
-    public void setAge(Integer age) {
-        this.age = age;
-    }
-}
