--- module12/src/main/java/ru/sberbank/edu/service/UserCache.java
@@ -1,26 +1,0 @@
-package ru.sberbank.edu.service;
-
-import org.springframework.stereotype.Component;
-
-import javax.annotation.PostConstruct;
-import java.util.HashMap;
-import java.util.Map;
-
-@Component
-public class UserCache {
-    private Map<String, UserInfo> cache = new HashMap<>();
-
-    public UserInfo get(String id) {
-        return cache.get(id);
-    }
-
-    private void create(UserInfo info) {
-        cache.put(info.getLogin(), info);
-
-    }
-
-    @PostConstruct
-    public void init() {
-        create(new UserInfo("admin", "123456", "Одмен", "ADMIN"));
-    }
-}
