+++ module13/docker/Dockerfile
@@ -1,0 +1,4 @@
+FROM openjdk:17
+WORKDIR /app
+COPY springRest-1.0-SNAPSHOT.jar /app
+ENTRYPOINT ["java", "-jar", "/app/springRest-1.0-SNAPSHOT.jar"]
