+++ module13/docker/image.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+docker build -t user_service_image .
+#docker images
+#docker-compose up
