+++ module13/src/main/resources/application-h2.yml
@@ -1,0 +1,40 @@
+logging:
+  level:
+    root: INFO
+
+    org:
+      springframework:
+        boot:
+          autoconfigure: INFO
+
+server:
+  port: 8082
+  address: localhost
+
+spring:
+  datasource:
+    url: jdbc:h2:mem:testdb
+    #    url: jdbc:h2:file:./app.db;AUTO_SERVER=TRUE
+    driverClassName: org.h2.Driver
+    username: sa
+    password: 123456
+  jpa:
+    show_sql: true
+    database-platform: org.hibernate.dialect.H2Dialect
+    hibernate:
+      ddl-auto: create-drop
+    defer-datasource-initialization: true
+    properties:
+      hibernate:
+        show_sql: true
+        format_sql: true
+  sql:
+    init:
+      mode: always
+
+  h2:
+    console:
+      enabled: true
+      path: /h2-console
+      settings.trace: false
+      settings.web-allow-others: false
