+++ module13/src/main/resources/application-postgres.yml
@@ -1,0 +1,5 @@
+spring:
+  datasource:
+    url: jdbc:postgresql://localhost:8089/postgres
+    username: sa
+    password: 123456
